moving tests to assets
